speaksite/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
